+++ pkg/analyzer/CHANGELOG.md
@@ -9,6 +9,7 @@
 * Removed `NullSafetyUnderstandingFlag`.
 * Removed `functionTypeAliasElement` from `TypeSystem.instantiateToBounds2`.
 * Added `Annotation.typeArguments` in preparation for supporting #44838.
+* Removed `actualUri` from `UriResolver.resolveAbsolute`.
 
 ## 0.41.1
 * Updated `PackageBuildWorkspace` that supports `package:build` to stop